Initial check-in of Franz' CLIM code
The intent is to be able to compile it with Clozure CL on all
platforms supported by Clozure CL.

Initially, it will use the CLX backend via the Lisp-only XLIB library
as a starting point to get it up and running on Linux.

I have removed the Franz' foreign function based XLIB and replaced
it with a Lisp based XLIB.

Eventually, we will want a GTK backend for Linux, a Cocoa backend for
Mac OSX, and a Win32/Win64 backend for Windows.  And appropriate backends
for any other supported platforms.